split the binary evasion detectors
