+++ art/defences/detector/evasion/__init__.py
@@ -2,5 +2,6 @@
 Module implementing detector-based defences against evasion attacks.
 """
 from art.defences.detector.evasion.evasion_detector import EvasionDetector
-from art.defences.detector.evasion.binary_detector import BinaryInputDetector, BinaryActivationDetector
+from art.defences.detector.evasion.binary_input_detector import BinaryInputDetector
+from art.defences.detector.evasion.binary_activation_detector import BinaryActivationDetector
 from art.defences.detector.evasion.subsetscanning.detector import SubsetScanningDetector
